Read Registry for kstat tunables
Upon Driver load read the Registry
"\HKEY_LOCAL_MACHINE\SYSTEM\ControlSet001\Services\ZFSin"

and set values for the entries that correspond to a kstat name. This
allows kstat tunable to be set for boot and reboot.

The Registry entry "hostid" is automatically created if not set.

With the exception of "hostid" (32bit, REG_DWORD) all kstat/Registry
tunables are 64bit / REG_QWORD.

For example:

"hostid"      REG_DWORD "0x12345678"
"zfs_arc_max" REG_QWORD "0x80000000"

Would set the ARC maximum size to 2GB of memory (2 * 1024 * 1024 * 1024)

+++ ZFSin/spl/module/spl/spl-windows.c
@@ -75,7 +75,7 @@ extern uint64_t		segkmem_total_mem_allocated;
 #define MAXHOSTNAMELEN 64
 extern char hostname[MAXHOSTNAMELEN];
 
-unsigned long spl_hostid = 0;
+uint32_t spl_hostid = 0;
 
 /*
  * Solaris delay is in ticks (hz) and Windows in 100 nanosecs
